modularize scripts
Changed sampling. I found a small error -there are duplicate indices in the data- that lead to a wrong behavior. Fixed it. Also, it was sampling all the patches in a song. So, batch size was not n. Changed it as well. Modularized "build iteratively". Functions are in module.py. There was a mistake in JP data. You need to run correct_JP_data.py. That script drops a song that has somehow two parts in two different places in the data. Also cleaned unnecessary files.